Add SrhRepo interface for search feature
This commit adds the SrhRepo interface for the search feature. The interface defines a method for searching food items based on a search text. The method returns a list of FoodEntity objects wrapped in an Either monad, which can represent either a success or a failure. The implementation of this interface will be done in a later commit.