Add SrhFoodUseCase to search domain usecases
This commit adds the SrhFoodUseCase to the search domain usecases. The use case is responsible for searching for food entities based on a search query. It takes in a string parameter and returns a list of FoodEntity objects wrapped in an Either monad. The SrhFoodUseCase implements the UseCase abstract class and uses the SrhRepo interface to perform the search query.