Refactor search_cubit.dart to use Cloud Firestore for searching food items
This commit refactors the search_cubit.dart file to use Cloud Firestore for searching food items instead of the previous implementation. The srhFood method now returns a Stream of QuerySnapshot<Map<String, dynamic>> instead of a Future. This change was made to improve the performance and scalability of the search feature. Additionally, error handling was improved to catch any unexpected errors and emit a SearchFailure state with the error message.